Set ApplicationRoute's model to session
Since ApplicationRoute is the the top-most route that is activated when
the user first loads the application from whatever url, it makes sense
to set this route's model to session.

Session id is fetched from cookies. If it's empty, the session is set to
null and the user is considered not authenticated. If, on the other
hand, the session id is not empty, session data are requested from the
server via AJAX. If the session is found on the server, valid user is
returned to the application, and the session is considered
authenticated. If the session is not found, the server returns the 401
response and model initialization fails. Which is just the right point
to redirect to the login route.

Since Ember.js supports promises while initializing models, and in our
case a promise is returned, it does not proceed until a response from
the server has been received.