Implement session management and login
Add OpenCBS.Session as a singleton that is globally accessible from any
part of the application.

When the ApplicationRoute is entered the global session instance is
initialized by hitting the server and checking if the session is
authenticated or not. If not, the user is redirected to the login page
in the IndexRoute, which is a descendant of the AuthenticatedRoute.