Improve the alerts grid
As we want the header of the grid to be fixed at the top of the
container and not scroll with the content, we have to set bot the header
and content positioning to absolute. But this poses a slight problem:
since both the header and content columns' sizes are set in
percentages, their absolute widths will differ slightly because the
parent nodes for these elements differ in width -- header *does not*
have the scrollbar, whereas the content area has.

To overcome this difference we have to dynamically set the padding-right
of the header to the width of the scrollbar.